Incorporated required changes regarding class imports and Renamed decorators
