+++ a10_octavia/controller/worker/tasks/server_tasks.py
@@ -18,7 +18,7 @@ from oslo_config import cfg
 from taskflow import task
 
 from a10_octavia.controller.worker.tasks import utils
-from a10_octavia.controller.worker.tasks.decorator import axapi_client_decorator
+from a10_octavia.controller.worker.tasks.decorators import axapi_client_decorator
 
 CONF = cfg.CONF
 LOG = logging.getLogger(__name__)
